+++ compiler/rustc_ast_lowering/src/lib.rs
@@ -2276,7 +2276,8 @@ impl<'hir> LoweringContext<'_, 'hir> {
         let hir_id = self.lower_node_id(param.id);
         let param_attrs = &param.attrs;
         let param_span = param.span();
-        let param = hir::GenericParam {
+        self.lower_attrs(hir_id, param_attrs, param_span, Target::from_generic_param(&param));
+        hir::GenericParam {
             hir_id,
             def_id: self.local_def_id(param.id),
             name,
@@ -2285,9 +2286,7 @@ impl<'hir> LoweringContext<'_, 'hir> {
             kind,
             colon_span: param.colon_span.map(|s| self.lower_span(s)),
             source,
-        };
-        self.lower_attrs(hir_id, param_attrs, param_span, Target::from_generic_param(&param));
-        param
+        }
     }
 
     fn lower_generic_param_kind(
