+++ compiler/rustc_attr_ir/src/canonical_symbols.rs
