+++ compiler/rustc_attr_ir/src/lib.rs
@@ -4,22 +4,42 @@
 //! For detailed documentation about attribute processing,
 //! see [rustc_attr_parsing](https://doc.rust-lang.org/nightly/nightly-rustc/rustc_attr_parsing/index.html).
 
+// tidy-alphabetical-start
+#![feature(const_default)]
+#![feature(const_trait_impl)]
+#![feature(default_field_values)]
+#![feature(derive_const)]
+#![feature(exhaustive_patterns)]
+#![feature(variant_count)]
+#![recursion_limit = "256"]
+// tidy-alphabetical-end
+
+pub use attr::*;
+pub use canonical_symbols::*;
 pub use data_structures::*;
 pub use encode_cross_crate::EncodeCrossCrate;
+pub use lang_items::*;
 pub use pretty_printing::PrintAttribute;
-
+pub use stability::*;
+pub use target::{MethodKind, Target};
+pub mod attr;
+pub mod canonical_symbols;
 mod data_structures;
 pub mod diagnostic;
+pub mod diagnostic_items;
 mod encode_cross_crate;
+pub mod lang_items;
 mod pretty_printing;
-
+pub mod stability;
+pub mod target;
+pub mod weak_lang_items;
 /// A trait for types that can provide a list of attributes given a `TyCtxt`.
 ///
 /// It allows `find_attr!` to accept either a `DefId`, `LocalDefId`, `OwnerId`, or `HirId`.
 /// It is defined here with a generic `Tcx` because `rustc_hir` can't depend on `rustc_middle`.
 /// The concrete implementations are in `rustc_middle`.
 pub trait HasAttrs<'tcx, Tcx> {
-    fn get_attrs(self, tcx: &Tcx) -> &'tcx [crate::Attribute];
+    fn get_attrs(self, tcx: &Tcx) -> &'tcx [crate::attr::Attribute];
 }
 
 /// Finds attributes in sequences of attributes by pattern matching.
@@ -71,7 +91,7 @@ macro_rules! find_attr {
     };
     ($tcx: expr, $id: expr, $pattern: pat $(if $guard: expr)? => $e: expr) => {{
         $crate::find_attr!(
-            $crate::attrs::HasAttrs::get_attrs($id, &$tcx),
+            $crate::HasAttrs::get_attrs($id, &$tcx),
             $pattern $(if $guard)? => $e
         )
     }};
@@ -85,13 +105,13 @@ macro_rules! find_attr {
         'done: {
             for i in $attributes_list {
                 #[allow(unused_imports)]
-                use rustc_hir::attrs::AttributeKind::*;
-                let i: &rustc_hir::Attribute = i;
+                use $crate::AttributeKind::*;
+                let i: &$crate::Attribute = i;
                 match i {
-                    rustc_hir::Attribute::Parsed($pattern) $(if $guard)? => {
+                    $crate::Attribute::Parsed($pattern) $(if $guard)? => {
                         break 'done Some($e);
                     }
-                    rustc_hir::Attribute::Unparsed(..) => {}
+                    $crate::Attribute::Unparsed(..) => {}
                     // In lint emitting, there's a specific exception for this warning.
                     // It's not usually emitted from inside macros from other crates
                     // (see https://github.com/rust-lang/rust/issues/110613)
