+++ compiler/rustc_attr_ir/src/pretty_printing.rs
