+++ compiler/rustc_attr_ir/src/target.rs
@@ -5,10 +5,6 @@ use std::fmt::{self, Display};
 use rustc_ast::visit::AssocCtxt;
 use rustc_ast::{AssocItemKind, ForeignItemKind, ast};
 use rustc_macros::StableHash;
-
-use crate::def::DefKind;
-use crate::{Item, ItemKind, TraitItem, TraitItemKind, hir};
-
 #[derive(Copy, Clone, PartialEq, Debug, Eq, StableHash)]
 pub enum GenericParamKind {
     Type,
@@ -125,50 +121,6 @@ impl Target {
         }
     }
 
-    pub fn from_item(item: &Item<'_>) -> Target {
-        match item.kind {
-            ItemKind::ExternCrate(..) => Target::ExternCrate,
-            ItemKind::Use(..) => Target::Use,
-            ItemKind::Static { .. } => Target::Static,
-            ItemKind::Const(..) => Target::Const,
-            ItemKind::Fn { .. } => Target::Fn,
-            ItemKind::Macro(..) => Target::MacroDef,
-            ItemKind::Mod(..) => Target::Mod,
-            ItemKind::ForeignMod { .. } => Target::ForeignMod,
-            ItemKind::GlobalAsm { .. } => Target::GlobalAsm,
-            ItemKind::TyAlias(..) => Target::TyAlias,
-            ItemKind::Enum(..) => Target::Enum,
-            ItemKind::Struct(..) => Target::Struct,
-            ItemKind::Union(..) => Target::Union,
-            ItemKind::Trait { .. } => Target::Trait,
-            ItemKind::TraitAlias(..) => Target::TraitAlias,
-            ItemKind::Impl(imp_) => Target::Impl { of_trait: imp_.of_trait.is_some() },
-        }
-    }
-
-    // FIXME: For now, should only be used with def_kinds from ItemIds
-    pub fn from_def_kind(def_kind: DefKind) -> Target {
-        match def_kind {
-            DefKind::ExternCrate => Target::ExternCrate,
-            DefKind::Use => Target::Use,
-            DefKind::Static { .. } => Target::Static,
-            DefKind::Const { .. } => Target::Const,
-            DefKind::Fn => Target::Fn,
-            DefKind::Macro(..) => Target::MacroDef,
-            DefKind::Mod => Target::Mod,
-            DefKind::ForeignMod => Target::ForeignMod,
-            DefKind::GlobalAsm => Target::GlobalAsm,
-            DefKind::TyAlias => Target::TyAlias,
-            DefKind::Enum => Target::Enum,
-            DefKind::Struct => Target::Struct,
-            DefKind::Union => Target::Union,
-            DefKind::Trait => Target::Trait,
-            DefKind::TraitAlias => Target::TraitAlias,
-            DefKind::Impl { of_trait } => Target::Impl { of_trait },
-            _ => panic!("impossible case reached"),
-        }
-    }
-
     pub fn from_ast_item(item: &ast::Item) -> Target {
         match item.kind {
             ast::ItemKind::ExternCrate(..) => Target::ExternCrate,
@@ -194,52 +146,17 @@ impl Target {
         }
     }
 
-    pub fn from_foreign_item_kind(kind: &ast::ForeignItemKind) -> Target {
-        match kind {
-            ForeignItemKind::Static(_) => Target::ForeignStatic,
-            ForeignItemKind::Fn(_) => Target::ForeignFn,
-            ForeignItemKind::TyAlias(_) => Target::ForeignTy,
-            ForeignItemKind::MacCall(_) => Target::MacroCall,
+    pub fn from_expr(expr: &ast::Expr) -> Self {
+        match &expr.kind {
+            ast::ExprKind::Closure(..) | ast::ExprKind::Gen(..) => Self::Closure,
+            ast::ExprKind::Paren(e) => Self::from_expr(&e),
+            ast::ExprKind::ForLoop { .. } => Self::ForLoop,
+            ast::ExprKind::Loop(..) => Self::Loop,
+            ast::ExprKind::While(..) => Self::While,
+            ast::ExprKind::Break(..) => Self::Break,
+            _ => Self::Expression,
         }
     }
-
-    pub fn from_trait_item(trait_item: &TraitItem<'_>) -> Target {
-        match trait_item.kind {
-            TraitItemKind::Const(..) => Target::AssocConst,
-            TraitItemKind::Fn(_, hir::TraitFn::Required(_)) => {
-                Target::Method(MethodKind::Trait { body: false })
-            }
-            TraitItemKind::Fn(_, hir::TraitFn::Provided(_)) => {
-                Target::Method(MethodKind::Trait { body: true })
-            }
-            TraitItemKind::Type(..) => Target::AssocTy,
-        }
-    }
-
-    pub fn from_foreign_item(foreign_item: &hir::ForeignItem<'_>) -> Target {
-        match foreign_item.kind {
-            hir::ForeignItemKind::Fn(..) => Target::ForeignFn,
-            hir::ForeignItemKind::Static(..) => Target::ForeignStatic,
-            hir::ForeignItemKind::Type => Target::ForeignTy,
-        }
-    }
-
-    pub fn from_generic_param(generic_param: &hir::GenericParam<'_>) -> Target {
-        match generic_param.kind {
-            hir::GenericParamKind::Type { default, .. } => Target::GenericParam {
-                kind: GenericParamKind::Type,
-                has_default: default.is_some(),
-            },
-            hir::GenericParamKind::Lifetime { .. } => {
-                Target::GenericParam { kind: GenericParamKind::Lifetime, has_default: false }
-            }
-            hir::GenericParamKind::Const { default, .. } => Target::GenericParam {
-                kind: GenericParamKind::Const,
-                has_default: default.is_some(),
-            },
-        }
-    }
-
     pub fn from_assoc_item_kind(kind: &ast::AssocItemKind, assoc_ctxt: AssocCtxt) -> Target {
         match kind {
             AssocItemKind::Const(_) => Target::AssocConst,
@@ -259,19 +176,29 @@ impl Target {
             AssocItemKind::MacCall(_) => Target::MacroCall,
         }
     }
-
-    pub fn from_expr(expr: &ast::Expr) -> Self {
-        match &expr.kind {
-            ast::ExprKind::Closure(..) | ast::ExprKind::Gen(..) => Self::Closure,
-            ast::ExprKind::Paren(e) => Self::from_expr(&e),
-            ast::ExprKind::ForLoop { .. } => Self::ForLoop,
-            ast::ExprKind::Loop(..) => Self::Loop,
-            ast::ExprKind::While(..) => Self::While,
-            ast::ExprKind::Break(..) => Self::Break,
-            _ => Self::Expression,
+    pub fn from_foreign_item_kind(kind: &ast::ForeignItemKind) -> Target {
+        match kind {
+            ForeignItemKind::Static(_) => Target::ForeignStatic,
+            ForeignItemKind::Fn(_) => Target::ForeignFn,
+            ForeignItemKind::TyAlias(_) => Target::ForeignTy,
+            ForeignItemKind::MacCall(_) => Target::MacroCall,
         }
     }
-
+    pub fn from_generic_param(generic_param: &ast::GenericParam) -> Target {
+        match &generic_param.kind {
+            ast::GenericParamKind::Type { default, .. } => Target::GenericParam {
+                kind: GenericParamKind::Type,
+                has_default: default.is_some(),
+            },
+            ast::GenericParamKind::Lifetime { .. } => {
+                Target::GenericParam { kind: GenericParamKind::Lifetime, has_default: false }
+            }
+            ast::GenericParamKind::Const { default, .. } => Target::GenericParam {
+                kind: GenericParamKind::Const,
+                has_default: default.is_some(),
+            },
+        }
+    }
     pub fn name(self) -> &'static str {
         match self {
             Target::ExternCrate => "extern crate",
