+++ compiler/rustc_hir/src/intravisit.rs
@@ -66,6 +66,7 @@
 
 use rustc_ast::Label;
 use rustc_ast::visit::{VisitorResult, try_visit, visit_opt, walk_list};
+use rustc_attr_ir::attr::Attribute;
 use rustc_span::def_id::LocalDefId;
 use rustc_span::{Ident, Span, Symbol};
